[feat]OldCrow
보스몬스터 구현 시작.
->데스도어의 보스 몬스터 중 회색 까마귀 선택.
-> 스킬, 상태 등 토대가 되는 클래스 작성.
-> 애니메이션 삭제
기존에 있던 애니메이션을 삭제하고 새로운 애니메이션으로 변경

+++ Assets/Data/Data.cs
@@ -14,7 +14,7 @@ public class _Boss
 {
     public int BossId { get; set; }
     public string BossName { get; set;}
-    public int BossHp { get; set;}
+    public float BossHp { get; set;}
     public float BossAttackPower { get;set;}
     public float BossSpeed { get; set;}
 }
